+++ src/java/org/apache/cassandra/service/accord/AccordJournalValueSerializers.java
@@ -35,8 +35,8 @@ import org.apache.cassandra.net.MessagingService;
 import org.apache.cassandra.service.accord.serializers.CommandStoreSerializers;
 import org.apache.cassandra.service.accord.serializers.KeySerializers;
 
+import static accord.api.Journal.Load.ALL;
 import static accord.local.CommandStores.RangesForEpoch;
-import static org.apache.cassandra.service.accord.SavedCommand.Load.ALL;
 
 // TODO (required): test with large collection values, and perhaps split out some fields if they have a tendency to grow larger
 // TODO (required): alert on metadata size
@@ -56,16 +56,16 @@ public class AccordJournalValueSerializers
     }
 
     public static class CommandDiffSerializer
-    implements FlyweightSerializer<SavedCommand.Writer, SavedCommand.Builder>
+    implements FlyweightSerializer<AccordJournal.Writer, AccordJournal.Builder>
     {
         @Override
-        public SavedCommand.Builder mergerFor(JournalKey journalKey)
+        public AccordJournal.Builder mergerFor(JournalKey journalKey)
         {
-            return new SavedCommand.Builder(journalKey.id, ALL);
+            return new AccordJournal.Builder(journalKey.id, ALL);
         }
 
         @Override
-        public void serialize(JournalKey key, SavedCommand.Writer writer, DataOutputPlus out, int userVersion)
+        public void serialize(JournalKey key, AccordJournal.Writer writer, DataOutputPlus out, int userVersion)
         {
             try
             {
@@ -78,13 +78,13 @@ public class AccordJournalValueSerializers
         }
 
         @Override
-        public void reserialize(JournalKey key, SavedCommand.Builder from, DataOutputPlus out, int userVersion) throws IOException
+        public void reserialize(JournalKey key, AccordJournal.Builder from, DataOutputPlus out, int userVersion) throws IOException
         {
             from.serialize(out, userVersion);
         }
 
         @Override
-        public void deserialize(JournalKey journalKey, SavedCommand.Builder into, DataInputPlus in, int userVersion) throws IOException
+        public void deserialize(JournalKey journalKey, AccordJournal.Builder into, DataInputPlus in, int userVersion) throws IOException
         {
             into.deserializeNext(in, userVersion);
         }
