+++ src/java/org/apache/cassandra/service/accord/AccordSafeCommandStore.java
@@ -32,6 +32,7 @@ import accord.api.ProgressLog;
 import accord.api.RoutingKey;
 import accord.impl.AbstractSafeCommandStore;
 import accord.impl.CommandsSummary;
+import accord.local.CommandStores;
 import accord.local.NodeCommandStoreService;
 import accord.local.cfk.CommandsForKey;
 import accord.primitives.AbstractKeys;
