+++ test/distributed/org/apache/cassandra/service/accord/AccordJournalBurnTest.java
@@ -137,7 +137,7 @@ public class AccordJournalBurnTest extends BurnTestBase
                  operations,
                  10 + random.nextInt(30),
                  new RandomDelayQueue.Factory(random).get(),
-                 (node) -> {
+                 (node, agent) -> {
                      try
                      {
                          File directory = new File(Files.createTempDirectory(Integer.toString(counter.incrementAndGet())));
@@ -176,6 +176,4 @@ public class AccordJournalBurnTest extends BurnTestBase
             throw SimulationException.wrap(seed, t);
         }
     }
-
-
 }
