+++ test/unit/org/apache/cassandra/service/accord/AccordJournalOrderTest.java
@@ -89,7 +89,7 @@ public class AccordJournalOrderTest
         Runnable check = () -> {
             for (JournalKey key : res.keySet())
             {
-                SavedCommand.Builder diffs = accordJournal.loadDiffs(key.commandStoreId, key.id);
+                AccordJournal.Builder diffs = accordJournal.load(key.commandStoreId, key.id);
                 Assert.assertEquals(String.format("%d != %d for key %s", diffs.count(), res.get(key).intValue(), key),
                                     diffs.count(), res.get(key).intValue());
             }
